requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==0.5.2
antlr4-python3-runtime==4.8
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==23.3.0
braceexpand==0.1.7
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
defusedxml==0.7.1
-e git+https://github.com/facebookresearch/detectron2.git@94113be6e12db36b8c7601e13747587f19ec92fe#egg=detectron2
Django==4.2
django-forms-bootstrap==3.1.0
djangorestframework==3.14.0
easydict==1.9
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
hydra-core==1.1.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
jax==0.4.12
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
kornia==0.5.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.6.0
portalocker==2.7.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytorch-lightning==1.2.9
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.2.0
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
webdataset==0.2.48
Werkzeug==2.3.6
wldhx.yadisk-direct==0.0.6
wrapt==1.14.1
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
